chat: wrap the transcript incrementally and draw only the viewport
Every frame re-ran textwrap over every transcript item to place
thumbnails, deep-cloned the whole wrapped transcript into a Paragraph,
and any delta forced a full rewrap. Thumbnail rows were counted by a
separate routine that disagreed with rendering whenever a collapsed
notice or thinking header wrapped, so images landed on the wrong rows.
Scroll offsets were u16, so past 65,535 wrapped lines the viewport
start wrapped around.

The wrap cache now stores each item's row range under a (width,
verbose) key: in-place edits re-render only that item, removals
truncate, appends extend, and a key change rebuilds. Thumbnail slots
come from the same cached rows that are drawn. Rendering draws only the
visible slice, and offsets are usize throughout.

+++ crates/assistd/src/chat/app/tests.rs
@@ -298,7 +298,11 @@ fn mouse_wheel_scrolls_by_a_fixed_step() {
         (MouseEventKind::ScrollDown, MOUSE_WHEEL_STEP),
     ] {
         app.on_mouse(wheel(kind));
-        assert_eq!(app.output.scroll_offset(), expected, "after {kind:?}");
+        assert_eq!(
+            app.output.scroll_offset(),
+            usize::from(expected),
+            "after {kind:?}"
+        );
     }
 }
